+++ src/core/index.ts
@@ -1,5 +1,6 @@
 export * from './constants'
 export * from './colors'
+export * from './debug'
 export * from './geom'
 export * from './graphql'
 export * from './datetime'
